requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlite3
openai==1.3.0
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
httpx==0.27.0
# For testing
pytest==7.4.0
# For database migrations
alembic==1.12.0
# For database ORM
sqlalchemy==2.0.20
# For database schema generation
databases==0.6.2
# For database connection pooling
asyncpg==0.28.0
# For database management
psycopg2-binary==2.9.6
# For database migrations
migrate==0.13.0
# For database schema generation
tortoise-orm==0.21.0
# For database connection pooling
aiosqlite==0.17.0
# For database management
mysql-connector-python==8.0.34
# For database migrations
mariadb==1.0.10
# For database schema generation
pymysql==1.0.3
# For database connection pooling
pyodbc==4.0.39